fix: changes size_t to int and range in the exception of Form constructor.

+++ ex01/Bureaucrat.hpp
@@ -6,7 +6,7 @@
 /*   By: arabelo- <arabelo-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 02:25:44 by arabelo-          #+#    #+#             */
-/*   Updated: 2024/07/02 17:42:00 by arabelo-         ###   ########.fr       */
+/*   Updated: 2024/08/25 01:35:13 by arabelo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,14 +22,14 @@ class Form;
 class Bureaucrat {
 	private:
 		const std::string	_name;
-		size_t				_grade;
+		int					_grade;
 	public:
 		Bureaucrat(void);
-		Bureaucrat(std::string name, size_t grade);
+		Bureaucrat(std::string name, int grade);
 		Bureaucrat(const Bureaucrat &obj);
 		Bureaucrat	&operator=(const Bureaucrat &obj);
 		~Bureaucrat(void);
-		size_t		getGrade(void) const;
+		int			getGrade(void) const;
 		std::string	getName(void) const;
 		void		increment(void);
 		void		decrement(void);
